phase d'exploration

+++ manager/executeur_taches/include/FinalApproachingClient.h
@@ -13,6 +13,6 @@ class FinalApproachingClient{
 public :
 	FinalApproachingClient();
 	virtual  ~FinalApproachingClient();
-	void starting(int8_t machineType, int32_t machineSide);
+	void starting(int8_t machineType, int8_t machineSide);
 };
 #endif
